add user auth to the app

+++ src/app/routers/post.py
@@ -5,13 +5,14 @@ from sqlalchemy.orm import Session
 
 import models
 import schemas
-import views
-from app import database, utils
+from app import database, utils, security
 
 from config import FASTAPI_TUT_DATABASE_URL
 
 ### Initialization
 connection =  database.Connection(FASTAPI_TUT_DATABASE_URL)
+
+authorized_users = security.AuthorizedUser()
 
 router = APIRouter(
     prefix="/posts",
@@ -21,11 +22,11 @@ router = APIRouter(
 ### Operation on Posts
 
 @router.post("/", 
-          status_code=status.HTTP_201_CREATED, response_model=views.PostDisplay)
-def create_post(payload: schemas.Post, db: Session = Depends(connection)):  
-    
+          status_code=status.HTTP_201_CREATED, response_model=schemas.PostDisplayView)
+def create_post(payload: schemas.Post, current_user: schemas.TokenData = Depends(authorized_users), db: Session = Depends(connection)):  
+    new_post = payload.dict() | {"owner_id": current_user.user_id}
     results = database.createItem(
-        table=models.Post,item=payload.dict(),session=db
+        table=models.Post,item=new_post ,session=db
     )
     
     if not results:
@@ -35,10 +36,12 @@ def create_post(payload: schemas.Post, db: Session = Depends(connection)):
                 "msg": f"Post creation fail"
             }],           
         )
+    
+    # print(current_user)
         
     return results
 
-@router.get("/", response_model=List[views.PostDisplay])
+@router.get("/", response_model=List[schemas.PostDisplayView])
 def read_all_posts(db: Session = Depends(connection)):
     results = database.readAllItem(table=models.Post, session=db)
     
@@ -52,27 +55,45 @@ def read_all_posts(db: Session = Depends(connection)):
         )
     return results
     
-@router.get("/{post_id}", response_model=views.PostDisplay)
+@router.get("/{post_id}", response_model=schemas.PostDisplayView)
 def read_post(post_id: int, db: Session = Depends(connection)):
-    # validateAndGetPost(post_id)
-    results = database.readItemById(
-        table=models.Post, item_id=post_id, session=db
+    return __get_post(post_id, db)
+
+@router.get("/{user_id}", response_model=List[schemas.PostDisplayView])
+def read_post_by_user(user_id: int, db: Session = Depends(connection)):
+    results = database.readItems(
+        table=models.Post,
+        filters=[models.Post.owner_id == user_id],
+        fetching_all=True,
+        session=db
     )
     
     if not results:
-        raise utils.notFoundException(
-            msg=f"Post with post_id={post_id} does not exists or ID our of range"
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{
+                "msg": f"No post exist",
+                "data": results
+            }],           
         )
         
     return results
 
-@router.put("/likes/{post_id}", response_model=views.PostDisplay)
-def update_postLikes(post_id: int, dislike: bool = False, db: Session = Depends(connection)):
+@router.put("/likes/{post_id}", response_model=schemas.PostDisplayView)
+def update_postLikes(
+    post_id: int, dislike: bool = False, 
+    current_user: schemas.TokenData = Depends(authorized_users), db: Session = Depends(connection)
+):
+    post = __get_post(post_id, db)
+    
+    if post.owner_id != current_user.user_id: 
+        raise utils.ForbiddenException(msg="Not authorized to perform requested action")
+    
     results = database.updateItemById(**{
         "table": models.Post,
         "item_id": post_id,
         "set_values": {
-            "likes": models.Post.likes + (1 - dislike*2)
+            "likes": post.likes + (1 - dislike*2)
         }
     }, session=db)
 
@@ -84,7 +105,14 @@ def update_postLikes(post_id: int, dislike: bool = False, db: Session = Depends(
     return results
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_posts(post_id: int, db: Session = Depends(connection)):
+def delete_posts(post_id: int, 
+    current_user: schemas.TokenData = Depends(authorized_users), db: Session = Depends(connection)
+):
+    post = __get_post(post_id, db)
+    
+    if post.owner_id != current_user.user_id: 
+        raise utils.ForbiddenException(msg="Not authorized to perform requested action")
+    
     results = database.deleteItemById(
         table=models.Post,item_id=post_id, session=db
     )
@@ -95,4 +123,16 @@ def delete_posts(post_id: int, db: Session = Depends(connection)):
         )
     
     return 
+
+def __get_post(post_id: int, db: Session = Depends(connection)) -> schemas.PostDisplayView:
+    results = database.readItemById(
+        table=models.Post, item_id=post_id, session=db
+    )
+    
+    if not results:
+        raise utils.notFoundException(
+            msg=f"Post with post_id={post_id} does not exists or ID our of range"
+        )
+        
+    return results
